refactor: make firmware and SecurityManager[2] portable
