+++ packages/eslint-plugin/src/rules/no-forbidden-imports.ts
@@ -9,6 +9,7 @@ const whitelistedImports = [
 	// These are browser-compatible
 	"fflate",
 	"dayjs",
+	"nrf-intel-hex",
 	"alcalzone-shared/arrays",
 	"alcalzone-shared/async",
 	"alcalzone-shared/comparable",
